requirements.txt
fastapi
uvicorn
https
python-dotenv
pytest
pytest-asyncio
prometheus_client